requirements_hostel.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
pytest
httpx
python-dotenv
email-validator
